% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{bci.r}
\alias{bci.r}
\title{b and ci to r}
\usage{
bci.r(b, cil, ciu, n, k, result = c("cil", "ciu", "avg", "marge"))
}
\arguments{
\item{b}{the regression coefficient (untransformed only, see other functions e.g., expb.r() for transformed coefficients)}

\item{cil}{the lower bound of the 95 confidence interval}

\item{ciu}{the upper bound of the 95 confidence interval}

\item{n}{the sample size}

\item{k}{the total number of predictors, enter k = 1 if bivariate}
}
\description{
regression coefficient and confidence interval to correlation coefficient: b and ci -> t -> r\cr
unadjusted t-statistics estimate bivariate correlation coefficient (enter k = 1)\cr
adjusted t-statistics (e.g., from a regression) estimate partial correlation coefficient \href{https://www.journals.uchicago.edu/doi/abs/10.5243/jsswr.2013.24}{(aloe & thompson, 2013)}.
}
\examples{
bci.r(.156, .10, .225, 166, 3)
dat \%>\% mutate (r_from_bci = bci.r(reg_coef, reg_lowerci, reg_upperci, reg_n, reg_predictors)) -> dat

}
