% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{bse.r}
\alias{bse.r}
\title{b and se to r}
\usage{
bse.r(b, se, n, k)
}
\arguments{
\item{b}{the regression coefficient (un-transformed only, see other functions e.g., expb.r() for transformed coefficients)}

\item{se}{the standard error}

\item{n}{the analytical sample size}

\item{k}{the total number of predictors, enter k = 1 if bivariate}
}
\description{
regression coefficient and standard error to correlation coefficient\cr
unadjusted b and se estimate bivariate correlation coefficient (enter k = 1)\cr
adjusted b and se estimate partial correlation coefficient \href{https://www.journals.uchicago.edu/doi/abs/10.5243/jsswr.2013.24}{(aloe & thompson, 2013)}.
}
\examples{
bse.r(.25, .09, 100, 3)
dat \%>\% mutate (r_from_bse = bse.r(reg_coef, reg_se, reg_n, reg_predictors)) -> dat

}
