% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{bse.t}
\alias{bse.t}
\title{b and se to t}
\usage{
bse.t(b, se)
}
\arguments{
\item{b}{the regression coefficient (non-transformed only, see other functions for transformed regression coefficients)}

\item{se}{the standard error}
}
\description{
regression coefficient and standard error to t-statistics
}
\examples{
bse.t(.25, .09)
dat \%>\% mutate (t_from_b_se = bse.t(reg_coef, reg_se)) -> dat

}
