% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{ci.se}
\alias{ci.se}
\title{ci to se}
\usage{
ci.se(
  est,
  cil,
  ciu,
  n,
  k,
  result = c("ciu", "cil", "avg", "diff", "marge"),
  use_z = FALSE
)
}
\arguments{
\item{est}{the point estimate (e.g., a regression coefficient, a mean)}

\item{cil}{the lower bound of the confidence interval}

\item{ciu}{the upper bound of the confidence interval}

\item{n}{the sample size}

\item{k}{the total number of predictors. for converting the ci associated with a raw mean, enter k = 0.}

\item{result}{select one of: \itemize{
\item"ciu" will compute se based on upper bound of ci (this is the default if unspecified)
\item"cil" will compute se based on lower bound of ci
\item"avg" will compute se from both ciu and cil and take the average
\item"diff" will give the difference between the se computed using ciu and using cil. for transformed point estimate, see sediff() function.
}}

\item{use_z}{this is a logical T or F. If T, uses critical Z = 1.96 to compute SE; if F, then uses n and k to compute df}
}
\description{
compute standard error from confidence interval\cr
we recommend checking the computations by using result = "diff". see below.\cr
computations are based on t-distribution, not z\cr
}
\examples{
# e.g., regression b=.156, 95CI [.10, .225], n=166, 3 total predictors
ci.se(.156, .10, .225, 166, 3)

}
