% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{crosstab.prev}
\alias{crosstab.prev}
\title{computing prevalence of the outcome in the reference group}
\usage{
crosstab.prev(c, d)
}
\arguments{
\item{c}{# of non-exposed cases}

\item{d}{# of non-exposed non-cases}
}
\description{
computing prevalence of the outcome in the reference group
}
