% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{crosstab.rr}
\alias{crosstab.rr}
\title{risk ratio computation from crosstab}
\usage{
crosstab.rr(a, b, c, d)
}
\arguments{
\item{a}{# of exposed cases}

\item{b}{# of exposed non-cases}

\item{c}{# of non-exposed cases}

\item{d}{# of non-exposed non-cases}
}
\description{
risk ratio computation from crosstab
}
