% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{d.g}
\alias{d.g}
\title{d to g}
\usage{
d.g(d, n, type = c("independent", "paired"))
}
\arguments{
\item{d}{cohen's d}

\item{type}{type of cohen's d: independent or paired-sample}

\item{totn}{sample size (sum of two independent groups or total number of paired samples)}
}
\description{
convert cohen's d to hedges' g\cr
Equation 4.22 in Borenstein et al. (2009), Introduction to Meta-Analysis.
}
\examples{
d.g(.30, 68, type = "independent")
d.g(.30, 68, type = "paired")

}
