% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{d.r}
\alias{d.r}
\title{d to r}
\usage{
d.r(d, n1, n2, dir)
}
\arguments{
\item{d}{cohen's d}

\item{n1}{cell size of group 1 (if missing, defaults correction factor a = 4, assumes n1=n2)}

\item{n2}{cell size of group 2 (if missing, defaults correction factor a = 4, assumes n1=n2)}

\item{dir}{if cohen's d is in absolute value, provide the empirical direction:\cr
+1 for positive associaitons (e.g., group 1 - group 2 >= 0) and -1 for negative associations (e.g., group 1 - group 2 < 0)}

\item{k}{total number of predictors}
}
\description{
cohen's d to correlation coefficient
}
\examples{
d.t(.80, 15, 20, 1, -1)
dat \%>\% mutate (t_from_d = d.t(cohend, stress_n, control_n, numpred, emp_dir)) -> dat

}
