% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{d.t}
\alias{d.t}
\title{d to t}
\usage{
d.t(d, n1, n2, k, dir)
}
\arguments{
\item{d}{cohen's d}

\item{n1}{cell size of group 1}

\item{n2}{cell size of group 2}

\item{k}{total number of predictors}

\item{dir}{if cohen's d is in absolute value, provide the empirical direction:\cr
+1 for positive associaitons and -1 for negative associations}
}
\description{
cohen's d to t-statistics
}
\examples{
d.t(.80, 15, 20, 1, -1)
dat \%>\% mutate (t_from_d = d.t(cohend, stress_n, control_n, numpred, emp_dir)) -> dat

}
