% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{dept.d}
\alias{dept.d}
\title{Dependent-samples t to Cohen's d}
\usage{
dept.d(
  t,
  n,
  method = c("dz", "md"),
  sd_diff = NULL,
  sd_pre = NULL,
  sd_post = NULL,
  r = NULL
)
}
\arguments{
\item{t}{t statistic from paired-samples test}

\item{n}{number of pairs (sample size)}

\item{method}{either "dz" (default; simple t/sqrt(n)) 
or "md" (Morris & DeShon, 2002 adjustment)}

\item{sd_diff}{(optional) SD of difference scores. 
Used if method = "md" and provided directly.}

\item{sd_pre}{(optional) SD of pre-scores. Required for method="md"
if sd_diff is not supplied.}

\item{sd_post}{(optional) SD of post-scores. Required for method="md"
if sd_diff is not supplied.}

\item{r}{(optional) correlation between pre- and post-scores. Required 
for method="md" if sd_diff is not supplied.}
}
\value{
Cohen's d
}
\description{
Convert a paired-samples t statistic into Cohen's d.
}
\examples{
dept.d(t = 2.5, n = 30)  # simple dz
dept.d(t = 2.5, n = 30, method = "md", sd_pre = 10, sd_post = 10, r = 0.7)
dept.d(t = 2.5, n = 30, method = "md", sd_diff = 6)

}
