% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pustejovsky.R
\name{dicho.d.z}
\alias{dicho.d.z}
\title{cohen's d from dichotomizing/extreme group design to z}
\usage{
dicho.d.z(d, p1, p2, n1, n2, k, dir)
}
\arguments{
\item{d}{cohen's d}

\item{p1}{cutoff percentile for group 1 (sample-based, please see other formulas for population-based in pustejovsky, 2014)}

\item{p2}{cutoff percentile for group 2 (sample-based, please see other formulas for population-based in pustejovsky, 2014)}

\item{n1}{cell size of group 1}

\item{n2}{cell size of group 2}

\item{k}{kth term for taylor series approximation (enter 1 to 5)}

\item{dir}{if cohen's d is in absolute value, provide the empirical direction:\cr
+1 for positive associations (e.g., group 1 - group 2 >= 0) and -1 for negative associations (e.g., group 1 - group 2 < 0)}
}
\description{
cohen's d from dichotomizing/extreme group design to fisher's z\cr
see \href{https://psycnet.apa.org/buy/2013-34335-001}{pustejovsky, 2014. psychological methods.} \cr
returns z based on taylor series approximation
}
\examples{
dicho.d.z(.80, 1/3, 1/3, 15, 15, 5, -1)

}
