% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pustejovsky.R
\name{dicho.meanci.z}
\alias{dicho.meanci.z}
\title{means and ci from dichotomizing/extreme group design to z}
\usage{
dicho.meanci.z(
  m1,
  m2,
  cil1,
  ciu1,
  cil2,
  ciu2,
  n1,
  n2,
  k,
  p1,
  p2,
  result = c("sediff", "cil", "ciu")
)
}
\arguments{
\item{m1}{mean of group 1}

\item{m2}{mean of group 2}

\item{cil1}{lower bound of 95 confidence interval for group 1 mean}

\item{ciu1}{upper bound of 95 confidence interval for group 1 mean}

\item{cil2}{lower bound of 95 confidence interval for group 2 mean}

\item{ciu2}{upper bound of 95 confidence interval for group 2 mean}

\item{n1}{cell size of group 1}

\item{n2}{cell size of group 2}

\item{k}{number of predictors}

\item{p1}{cutoff percentile for group 1 (sample-based, please see other formulas for population-based in pustejovsky, 2014)}

\item{p2}{cutoff percentile for group 2 (sample-based, please see other formulas for population-based in pustejovsky, 2014)}
}
\description{
means and confidence interval from dichotomizing/extreme group design to fisher's z\cr
see \href{https://psycnet.apa.org/buy/2013-34335-001}{pustejovsky, 2014. psychological methods.} \cr
returns extreme group r (referred to as r (subscript eg) in pustejovsky paper) \cr
for taylor series approximation, defaults 5th term (use dicho.d.z() if you'd like to adjust to lower term)
}
\examples{
dicho.meanci.z(5, 3, 3, 6, 1, 4, 30, 31, 0, .33, .33)

}
