% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pustejovsky.R
\name{dicho.meanci.zvar}
\alias{dicho.meanci.zvar}
\title{variance of z converted from dichotomizing/extreme group design mean and ci}
\usage{
dicho.meanci.zvar(
  m1,
  m2,
  cil1,
  ciu1,
  cil2,
  ciu2,
  n1,
  n2,
  k,
  p1,
  p2,
  type = c("vd.ce", "vd.eg")
)
}
\arguments{
\item{m1}{mean of group 1}

\item{m2}{mean of group 2}

\item{cil1}{lower bound of 95 confidence interval for group 1 mean}

\item{ciu1}{upper bound of 95 confidence interval for group 1 mean}

\item{cil2}{lower bound of 95 confidence interval for group 2 mean}

\item{ciu2}{upper bound of 95 confidence interval for group 2 mean}

\item{n1}{cell size of group 1}

\item{n2}{cell size of group 2}

\item{k}{number of predictors}

\item{p1}{cutoff percentile for group 1 (sample-based, please see other formulas for population-based in pustejovsky, 2014)}

\item{p2}{cutoff percentile for group 2 (sample-based, please see other formulas for population-based in pustejovsky, 2014)}

\item{type}{select one from below, please read \href{https://psycnet.apa.org/buy/2013-34335-001}{pustejovsky, 2014. psychological methods.}for recommendations\cr
there are situations where the "controlled experiment" formulas are recommended even for stats derived from dichotomize/extreme group designs\cr
so please refer to his paper\itemize{
\item"vd.ce" compute variance of d using the controlled experiment formulas
\item"vd.eg" compute variance of d using the extreme group formulas}}
}
\description{
compute the variance of z from t-statistics derived from dichotomizing/extreme group design mean and ci\cr
see \href{https://psycnet.apa.org/buy/2013-34335-001}{pustejovsky, 2014. psychological methods.} \cr
}
\examples{
dicho.meanci.zvar(5, 3, 3, 6, 1, 4, 30, 31, 0, .33, .33, "vd.eg")

}
