% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pustejovsky.R
\name{dicho.meansd.z}
\alias{dicho.meansd.z}
\title{means and sd from dichotomizing/extreme group design to z}
\usage{
dicho.meansd.z(m1, m2, sd1, sd2, n1, n2, p1, p2)
}
\arguments{
\item{m1}{mean of group 1}

\item{m2}{mean of group 2}

\item{sd1}{standard deviation of group 1}

\item{sd2}{standard deviation of group 2}

\item{n1}{cell size of group 1}

\item{n2}{cell size of group 2}

\item{p1}{cutoff percentile for group 1 (sample-based, please see other formulas for population-based in pustejovsky, 2014)}

\item{p2}{cutoff percentile for group 2 (sample-based, please see other formulas for population-based in pustejovsky, 2014)}

\item{dir}{if cohen's d is in absolute value, provide the empirical direction:\cr
+1 for positive associations (e.g., group 1 - group 2 >= 0) and -1 for negative associations (e.g., group 1 - group 2 < 0)}
}
\description{
means and sd from dichotomizing/extreme group design to fisher's z\cr
see \href{https://psycnet.apa.org/buy/2013-34335-001}{pustejovsky, 2014. psychological methods.} \cr
returns extreme group r (referred to as r (subscript eg) in pustejovsky paper) \cr
for taylor series approximation, defaults 5th term (use dicho.d.z() if you'd like to adjust to lower term)
}
\examples{
dicho.meansd.z(15, 13, 1.4, 0.9, 15, 15, 1/3, 1/3)

}
