% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pustejovsky.R
\name{dicho.meansd.zvar}
\alias{dicho.meansd.zvar}
\title{variance of z converted from dichotomizing/extreme group design mean and sd}
\usage{
dicho.meansd.zvar(m1, m2, sd1, sd2, n1, n2, p1, p2, type = c("vd.ce", "vd.eg"))
}
\arguments{
\item{m1}{mean of group 1}

\item{m2}{mean of group 2}

\item{sd1}{standard deviation of group 1}

\item{sd2}{standard deviation of group 2}

\item{n1}{cell size of group 1}

\item{n2}{cell size of group 2}

\item{p1}{cutoff percentile for group 1 (sample-based, please see other formulas for population-based in pustejovsky, 2014)}

\item{p2}{cutoff percentile for group 2 (sample-based, please see other formulas for population-based in pustejovsky, 2014)}

\item{type}{select one (read \href{https://psycnet.apa.org/buy/2013-34335-001}{pustejovsky, 2014. psychological methods.} for recommendations\cr
there are situations where the "controlled experiment" formulas are recommended even for stats derived from dichotomize/extreme group designs\cr
so please refer to his paper)\itemize{
\item"vd.ce" compute variance of d using the controlled experiment formulas
\item"vd.eg" compute variance of d using the extreme group formulas}}
}
\description{
compute the variance of z from means and sds derived from dichotomizing/extreme group design d\cr
see \href{https://psycnet.apa.org/buy/2013-34335-001}{pustejovsky, 2014. psychological methods.} \cr
}
\examples{
dicho.meansd.zvar(15, 13, 1.4, 0.9, 15, 15, 1/3, 1/3, "vd.eg")

}
