% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pustejovsky.R
\name{dicho.meanse.r}
\alias{dicho.meanse.r}
\title{means and se from dichotomizing/extreme group design to r}
\usage{
dicho.meanse.r(m1, m2, se1, se2, n1, n2, p1, p2)
}
\arguments{
\item{m1}{mean of group 1}

\item{m2}{mean of group 2}

\item{se1}{standard error of group 1}

\item{se2}{standard error of group 2}

\item{n1}{cell size of group 1}

\item{n2}{cell size of group 2}

\item{p1}{cutoff percentile for group 1 (sample-based, please see other formulas for population-based in pustejovsky, 2014)}

\item{p2}{cutoff percentile for group 2 (sample-based, please see other formulas for population-based in pustejovsky, 2014)}
}
\description{
means and standard errors from dichotomizing/extreme group design to correlation coefficient\cr
see \href{https://psycnet.apa.org/buy/2013-34335-001}{pustejovsky, 2014. psychological methods.} \cr
returns extreme group r (referred to as r subscript eg in pustejovsky paper)
}
\examples{
dicho.meanse.r(1.1, 0.5, 0.6, 0.2, 10, 18, .25, .25)

}
