% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pustejovsky.R
\name{dicho.meanse.zvar}
\alias{dicho.meanse.zvar}
\title{variance of z converted from dichotomizing/extreme group design mean and se}
\usage{
dicho.meanse.zvar(m1, m2, se1, se2, n1, n2, p1, p2, type = c("vd.ce", "vd.eg"))
}
\arguments{
\item{m1}{mean of group 1}

\item{m2}{mean of group 2}

\item{se1}{standard error of group 1}

\item{se2}{standard error of group 2}

\item{n1}{cell size of group 1}

\item{n2}{cell size of group 2}

\item{p1}{cutoff percentile for group 1 (sample-based, please see other formulas for population-based in pustejovsky, 2014)}

\item{p2}{cutoff percentile for group 2 (sample-based, please see other formulas for population-based in pustejovsky, 2014)}

\item{type}{select one from below, please read \href{https://psycnet.apa.org/buy/2013-34335-001}{pustejovsky, 2014. psychological methods.}for recommendations\cr
there are situations where the "controlled experiment" formulas are recommended even for stats derived from dichotomize/extreme group designs\cr
so please refer to his paper\itemize{
\item"vd.ce" compute variance of d using the controlled experiment formulas
\item"vd.eg" compute variance of d using the extreme group formulas}}
}
\description{
compute the variance of z from t-statistics derived from dichotomizing/extreme group design mean and se\cr
see \href{https://psycnet.apa.org/buy/2013-34335-001}{pustejovsky, 2014. psychological methods.} \cr
}
\examples{
dicho.meanse.z(1.1, 0.5, 0.6, 0.2, 10, 18, .25, .25, "vd.ce")

}
