% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{expb.r}
\alias{expb.r}
\title{exponentiated b and ci to r}
\usage{
expb.r(b, cil, ciu, n, k, use_z = F)
}
\arguments{
\item{b}{exponentiated logistic regression coefficient}

\item{cil}{the lower bound of the 95 confidence interval}

\item{ciu}{the upper bound of the 95 confidence interval}

\item{n}{the sample size}

\item{k}{the total number of predictors, enter k = 1 if bivariate}
}
\description{
exponentiated logistic regression coefficient and confidence interval to correlation coefficient\cr
expnentiated b and ci -> t -> r\cr
unadjusted t-statistics estimate bivariate correlation coefficient (enter k = 1)\cr
adjusted t-statistics (e.g., from a regression) estimate partial correlation coefficient \href{https://www.journals.uchicago.edu/doi/abs/10.5243/jsswr.2013.24}{(aloe & thompson, 2013)}.
}
\examples{
expb.r(4.05, 1.56, 10.51, 141, 8)
dat \%>\% mutate (r_from_expbci = expb.r(expb_coef, expb_lowerci, expb_upperci, expb_n, expb_predictors)) -> dat

}
