% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{f.r}
\alias{f.r}
\title{f to r}
\usage{
f.r(f, df, dir)
}
\arguments{
\item{f}{the f-statistics}

\item{df}{the degrees of freedom2}

\item{dir}{empirical direction: +1 or -1. articles sometimes report f-statistics in absolute value, may be prudent to check when extracting stats\cr
if the reported f-statistics is positive, read text to ensure it reflects a positive association, enter the f as is and enter dir = 1 if so\cr
if the reported f-statistics is positive, but text suggests a negative association, enter the f as is and enter dir = -1\cr
if the reported f-statistics is negative, you can simply enter the negative f-value and enter dir = 1\cr}
}
\description{
f-statistics (df1=1) to correlation coefficient\cr
unadjusted statistics only
}
\examples{
f.r(3.21, 97, -1)
dat \%>\% mutate (r_from_f = f.r(reg_f, reg_df, reg_emp_direction)) -> dat

}
