% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{freq.or}
\alias{freq.or}
\title{frequency to odds ratio}
\usage{
freq.or(a, b, c, d)
}
\arguments{
\item{a}{exposed, outcome}

\item{b}{exposed, no outcome}

\item{c}{unexposed, outcome}

\item{d}{unexposed, no outcome}
}
\description{
frequency to odds ratio
}
