% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{freq.varlogor}
\alias{freq.varlogor}
\title{variance of log OR}
\usage{
freq.varlogor(a, b, c, d, result = c("se", "var"))
}
\arguments{
\item{a}{exposed, outcome}

\item{b}{exposed, no outcome}

\item{c}{unexposed, outcome}

\item{d}{unexposed, no outcome}
}
\description{
variance of log OR
}
