% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{geo.d}
\alias{geo.d}
\title{geometric mean and ci to d}
\usage{
geo.d(
  m1,
  m2,
  cil1,
  ciu1,
  cil2,
  ciu2,
  n1,
  n2,
  k,
  result = c("sediff", "cil", "ciu")
)
}
\arguments{
\item{m1}{geometric mean of group 1}

\item{m2}{geometric mean of group 2}

\item{cil1}{lower bound of 95 confidence interval}

\item{ciu1}{upper bound of 95 confidence interval}

\item{cil2}{lower bound of 95 confidence interval}

\item{ciu2}{upper bound of 95 confidence interval}

\item{n1}{cell size of group 1}

\item{n2}{cell size of group 2}

\item{k}{total number of predictors (k=0 for raw)}
}
\description{
geometric means and confidence interval of two groups to cohen's d
}
\examples{
geo.d(.76, .75, .41, 1.38, .45, 1.26, 57, 355, 0)


}
