% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{geoci.se}
\alias{geoci.se}
\title{geometric mean and ci to se}
\usage{
geoci.se(geomean, cil, ciu, n, k, result = c("cil", "ciu"))
}
\arguments{
\item{geomean}{geometric mean}

\item{cil}{lower bound of 95 confidence interval}

\item{ciu}{upper bound of 95 confidence interval}

\item{n}{cell size}

\item{k}{total number of predictors (k=0 for raw)}
}
\description{
confidence interval around geometric mean to standard error of arithmetic mean of logged raw units
}
\examples{
geoci.se(.76, .41, 1.38, 57, 0)

}
