% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{lpm.var}
\alias{lpm.var}
\title{variance of the log odds ratio from a linear probability model coefficient}
\usage{
lpm.var(b, se_b, p0, delta = 1, results = c("se", "var"))
}
\arguments{
\item{b}{linear probability model coefficient (proportion scale)}

\item{se_b}{standard error of b, same scale}

\item{p0}{baseline prevalence in the reference group (proportion)}

\item{delta}{contrast width; 1 for a categorical contrast}
}
\description{
variance of the log odds ratio from a linear probability model coefficient
}
