% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{meanci.d}
\alias{meanci.d}
\title{mean and ci to d}
\usage{
meanci.d(m1, m2, cil1, ciu1, cil2, ciu2, n1, n2, k)
}
\arguments{
\item{m1}{mean of group 1}

\item{m2}{mean of group 2}

\item{n1}{cell size of group 1}

\item{n2}{cell size of group 2}

\item{k}{total number of predictors (k=0 for raw)}

\item{cil}{lower bound of 95 confidence interval}

\item{ciu}{upper bound of 95 confidence interval}
}
\description{
means and confidence interval of two groups to cohen's d
}
