% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{meansd.d}
\alias{meansd.d}
\title{mean and sd to d}
\usage{
meansd.d(m1, m2, sd1, sd2, n1, n2)
}
\arguments{
\item{m1}{mean of group 1}

\item{m2}{mean of group 2}

\item{sd1}{standard deviation of group 1}

\item{sd2}{standard deviation of group 2}

\item{n1}{cell size of group 1}

\item{n2}{cell size of group 2}
}
\description{
means and standard deviations of two groups to cohen's d
}
\examples{
meansd.d(15.1, 13.2, 2.3, 1.1, 25, 45)
dat \%>\% mutate (d_from_meansd = meansd.d(mean1, mean2, sd1, sd2, n1, n2)) -> dat

}
