% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{meanse.d}
\alias{meanse.d}
\title{mean and se to d}
\usage{
meanse.d(m1, m2, se1, se2, n1, n2)
}
\arguments{
\item{m1}{mean of group 1}

\item{m2}{mean of group 2}

\item{se1}{standard deviation of group 1}

\item{se2}{standard deviation of group 2}

\item{n1}{cell size of group 1}

\item{n2}{cell size of group 2}
}
\description{
means and standard errors of two groups to cohen's d
}
\examples{
meanse.d(15.1, 13.2, .64, .164, 25, 45)

}
