% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{meanse.r}
\alias{meanse.r}
\title{mean and se to r}
\usage{
meanse.r(m1, m2, se1, se2, n1, n2, k)
}
\arguments{
\item{m1}{mean of group 1}

\item{m2}{mean of group 2}

\item{se1}{standard error of group 1}

\item{se2}{standard error of group 2}

\item{n1}{cell size of group 1}

\item{n2}{cell size of group 2}

\item{k}{total number of predictors (k=0 for raw)}
}
\description{
means and standard error of two groups to correlation coefficient
}
\examples{
meanse.r(15.1, 13.2, .64, .164, 25, 45)
dat \%>\% mutate (r_from_meanse = meanse.r(stress_mean, control_mean, stress_se, control_se, stress_n, control_n)) -> dat

}
