% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{median.mean}
\alias{median.mean}
\title{median, minimum, and max to mean}
\usage{
\method{median}{mean}(median, min, max)
}
\arguments{
\item{median}{the median}

\item{min}{the minimum}

\item{max}{the maximum}
}
\description{
estimate mean from median, minimum and max. see \href{https://onlinelibrary.wiley.com/doi/full/10.4073/cmpn.2016.3}{polanin & snilstveit, 2016}
}
\examples{
median.mean(61, -11, 516)
dat \%>\% mutate (mean_from_medianminmax = median.mean(stress_median, stress_min, stress_max)) -> dat

}
