% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{median.r}
\alias{median.r}
\title{median, minimum, and max to r}
\usage{
\method{median}{r}(median1, median2, min1, max1, min2, max2, n1, n2)
}
\arguments{
\item{median1}{the median for group 1}

\item{median2}{the median for group 2}

\item{min1}{the minimum for group 1}

\item{max1}{the maximum for group 1}

\item{min2}{the minimum for group 2}

\item{max2}{the maximum for group 2}

\item{n1}{the cell size for group 1}

\item{n2}{the cell size for group 2}
}
\description{
estimate correlation coefficient from median, minimum and max. see \href{https://onlinelibrary.wiley.com/doi/full/10.4073/cmpn.2016.3}{polanin & snilstveit, 2016}
}
\examples{
median.r(61, 2, -11, 516, -10, 2159, 26, 52)
dat \%>\% mutate (r_from_medianminmax = median.r(stress_median, control_median, stress_minimum, control_minimum2, stress_maximum, control_maximum, stress_n, control_n)) -> dat


}
