% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{or.r}
\alias{or.r}
\title{or to r}
\usage{
or.r(or, n1, n2)
}
\arguments{
\item{or}{"true" odd ratio calculated from or = ad/bc (see description above)}

\item{n1}{cell size of group 1, if unknown, leave blank-- defaults correction factor a = 4 in estimation of d (assumes n1 = n2)}

\item{n2}{cell size of group 2, if unknown, leave blank-- defaults correction factor a = 4 in estimation of d (assumes n1 = n2)}
}
\description{
odd ratio to correlation coefficient: please only use this conversion for "true" odd ratios calculated by
or = ad/bc (a = # of exposed cases; b = # of exposed non-cases, c = # of unexposed cases, d = # of unexposed non-cases)
for estiamted or derived from exponentiating the coefficient of a logistic regression, use expb.r() instead
}
\examples{
or.r(1.25)
or.r(1.25, 25, 34)

}
