% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{orvar}
\alias{orvar}
\title{variance of or}
\usage{
orvar(a, b, c, d)
}
\arguments{
\item{or}{cohen's d estimate}

\item{n1}{cell size of group 1 for iv}

\item{n2}{cell size of group 2 for iv}
}
\description{
compute the sampling variance of or (true or only)
}
\examples{
orvar()

}
