% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{pairdiffsd.d}
\alias{pairdiffsd.d}
\title{paired groups (difference, SD) (based on Comprehensive Meta-analysis Software)}
\usage{
pairdiffsd.d(diff, sd_diff, r, totn)
}
\arguments{
\item{diff}{mean difference}

\item{sd_diff}{sd of difference scores}

\item{r}{pearson's r between the two conditions}

\item{totn}{total n}
}
\description{
mean difference scores and SD of difference scores --> cohen's d
}
\examples{
pairdiffsd.d (6, 12.430, .7, 86)
}
