% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{pairmeansd.d}
\alias{pairmeansd.d}
\title{paired groups mean and sd to d (based on Comprehensive Meta-analysis Software)}
\usage{
pairmeansd.d(m1, m2, sd1, sd2, r, totn)
}
\arguments{
\item{m1}{mean of condition 1}

\item{m2}{mean of condition 2}

\item{sd1}{standard deviation of condition 1}

\item{sd2}{standard deviation of condition 2}

\item{r}{pearson's r between the two conditions}

\item{totn}{total n}
}
\description{
means and standard deviations of paired groups to cohen's d
}
\examples{
pairmeansd.d(31.490, 24.630, 9.120, 10.220, .745, 41)

}
