% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{pairmeanst.d}
\alias{pairmeanst.d}
\title{Paired groups (means, t) (based on Comprehensive Meta-analysis Software)}
\usage{
pairmeanst.d(t, m1, m2, r, totn)
}
\arguments{
\item{t}{pairwise t-statistics}

\item{m1}{mean 1}

\item{m2}{mean 2}

\item{r}{pearson's r between the two conditions}

\item{totn}{total n}
}
\description{
t-statistics --> cohen's d with means to ascertain empirical direction
direction is based on m1 minus m2
}
\examples{
pairmeanst.d(5.788, 10, 9, .5, 115)
}
