% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{pairt.d}
\alias{pairt.d}
\title{Paired groups (N, t) (based on Comprehensive Meta-analysis Software)}
\usage{
pairt.d(t, r, totn, dir)
}
\arguments{
\item{t}{pairwise t-statistics}

\item{r}{pearson's r between the two conditions}

\item{totn}{total n}

\item{dir}{empirical direction of the effect (+1 or -1)}
}
\description{
t-statistics --> cohen's d
}
\examples{
pairt.d(3.270, .5, 36, 1)
}
