% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{pairvard}
\alias{pairvard}
\title{variance of d (paired-samples)}
\usage{
pairvard(d, r, totn)
}
\arguments{
\item{d}{cohen's d}

\item{r}{pearson's r between the paired measurements}

\item{totn}{sample size (number of paired observations)}
}
\description{
variance of cohen's d\cr
based on comprehensive meta-analysis software
}
\examples{
pairvard(.89, .70, 36)

}
