% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{pairvarg}
\alias{pairvarg}
\title{variance of g (paired-sample d)}
\usage{
pairvarg(d, r, totn)
}
\arguments{
\item{d}{cohen's d}

\item{totn}{sample size (total number of paired observations)}
}
\description{
variance of hedges' g\cr
based on comprehensive meta-analysis software
}
\examples{
pairvarg(.3, .7, 68)

}
