% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{percent.r}
\alias{percent.r}
\title{percent change to r}
\usage{
percent.r(percent, cil, ciu, n, k, result = c("cil", "ciu", "avg"))
}
\arguments{
\item{percent}{the percent change calculated from exponentiated logistic regression coefficient \cr
specifically computed such that percent change = 100*(exp(b) - 1); where b is the regression coefficient \cr
if percent change expressed in decimals, please *100 first \cr
or if percent change not computed with above equation, please compute back to exponentiated b and then use expb.r()}

\item{cil}{the lower bound of the confidence interval around the percent change}

\item{ciu}{the upper bound of the confidence interval around the percent change}

\item{n}{the sample size}

\item{k}{the total number of predictors, enter k = 1 if bivariate}

\item{result}{one of the following:\itemize{
\item"ciu" uses the upper bound of ci to compute se (defaults if missing input)
\item"cil" uses the lower bound of ci to compute se
\item"avg" uses both upper bound and lower bound and then take the average of the resulting se's
}}
}
\description{
percent change derived from transformed exponentiated regression coefficients and confidence interval to correlation coefficient\cr
percent change and ci -> expnentiated b and ci -> t -> r\cr
unadjusted t-statistics estimate bivariate correlation coefficient (enter k = 1)\cr
adjusted t-statistics (e.g., from a regression) estimate partial correlation coefficient \href{https://www.journals.uchicago.edu/doi/abs/10.5243/jsswr.2013.24}{(aloe & thompson, 2013)}.
}
\examples{
percent.r(7, 3, 11, 3134, 6)
dat \%>\% mutate (r_from_perc = percent.r(percent, percent_lowerci, percent_upperci, percent_n, percent_predictors)) -> dat

}
