% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{r.z}
\alias{r.z}
\title{r to z}
\usage{
r.z(r)
}
\arguments{
\item{r}{correlation coefficient}
}
\description{
convert correlation coefficient to fisher's z\cr
correlation coefficient can be either bivariate pearson's r or partial correlation coefficient (hedges & olkin, 1985)
}
\examples{
single value
r.z(.3)

dplyr examples: dataset named "dat" with a column of correlation coefficient values named "corr", create a new var called "z"
dat \%>\% mutate (z = r.z(corr)) -> dat

}
