% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{regp.r}
\alias{regp.r}
\title{p to r}
\usage{
regp.r(p, n, k, dir)
}
\arguments{
\item{p}{the p-value}

\item{n}{the sample size}

\item{k}{the number of predictors}

\item{dir}{empirical direction: +1 for positive associaitons and -1 for negative associations}
}
\description{
regression p-values (or other t distribution-based p-values) to correlation coefficient\cr
estimates t-statistics and then correlation coefficients\cr
}
\examples{
regp.r(.023, 100, 3, 1)
dat \%>\% mutate (r_from_pval = regp.r(reg_pval, reg_n, reg_predictors, reg_emp_direction)) -> dat

}
