% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{rr.r}
\alias{rr.r}
\title{risk ratio to r}
\usage{
rr.r(rr, prev, n1, n2)
}
\arguments{
\item{rr}{the risk ratio}

\item{prev}{the prevalence of the outcome in the reference group}

\item{n1}{the cell size of group 1 (if unknown, leave blank, assumes correcting factor a = 4 where n1 = n2)}

\item{n2}{the cell size of group 2 (if unknown, leave blank, assumes correcting factor a = 4 where n1 = n2)}
}
\description{
risk ratio to correlation coefficient
use crosstab.rr() to calculate rr and use crosstab.prev() to calculate prevalence of the outcome in reference group
then use the following to calculate rr to r
}
\examples{
rr.r(4.05, 1.56, 10.51, 141, 8)
dat \%>\% mutate (r_from_rr = rr.r(rr, rr_lowerci, rr_upperci, rr_n, rr_predictors)) -> dat

}
