% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{rvar}
\alias{rvar}
\title{variance of r}
\usage{
rvar(r, n)
}
\arguments{
\item{n}{sample size from which the r is derived}
}
\description{
compute the sampling variance of r (not recommended for meta)
}
\examples{
rvar(100)

}
