% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{t.d}
\alias{t.d}
\title{t to d}
\usage{
\method{t}{d}(t, n1, n2, dir, k = 1)
}
\arguments{
\item{t}{the independent t-statistics}

\item{n1}{group 1 sample size}

\item{n2}{group 2 sample size}

\item{dir}{empirical direction: +1 or -1. articles sometimes report t-statistics in absolute value, may be prudent to check when extracting stats\cr
if the reported t-statistics is positive, read text to ensure it reflects a positive association, enter the t as is and enter dir = 1 if so\cr
if the reported t-statistics is positive, but text suggests a negative association, enter the t as is and enter dir = -1\cr
if the reported t-statistics is negative, you can simply enter the negative t-value and enter dir = 1\cr
note: if your t-statistics is computed from taking the ratio of b and se, e.g., using bse.t(), then it should already have empirical direction "built-in" since b is directional itself, so simply enter dir = 1}
}
\description{
independent t-statistics to cohen's d\cr
}
\examples{
t.g(2.14, 20, 35, -1)
dat \%>\% mutate (g = t.g(t_stat, n1, n2, direction)) -> dat

}
