% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{t.r}
\alias{t.r}
\title{t to r}
\usage{
\method{t}{r}(t, n, k, dir)
}
\arguments{
\item{t}{the t-statistics}

\item{n}{the sample size}

\item{k}{the total number of predictors, enter k = 1 if bivariate}

\item{dir}{empirical direction: +1 or -1. articles sometimes report t-statistics in absolute value, may be prudent to check when extracting stats\cr
if the reported t-statistics is positive, read text to ensure it reflects a positive association, enter the t as is and enter dir = 1 if so\cr
if the reported t-statistics is positive, but text suggests a negative association, enter the t as is and enter dir = -1\cr
if the reported t-statistics is negative, you can simply enter the negative t-value and enter dir = 1\cr
note: if your t-statistics is computed from taking the ratio of b and se, e.g., using bse.t(), then it should already have empirical direction "built-in" since b is directional itself, so simply enter dir = 1}
}
\description{
t-statistics to correlation coefficient\cr
unadjusted t-statistics estimate bivariate correlation coefficient (enter k = 1)\cr
adjusted t-statistics (e.g., from a regression) estimate partial correlation coefficient \href{https://www.journals.uchicago.edu/doi/abs/10.5243/jsswr.2013.24}{(aloe & thompson, 2013)}.
}
\examples{
t.r(2.14, 100, -1)
dat \%>\% mutate (r_from_t = t.r(t_stat, analytical_n, direction)) -> dat

}
