% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{varg}
\alias{varg}
\title{variance of g}
\usage{
varg(d, n1, n2)
}
\arguments{
\item{d}{cohen's d}

\item{n1}{number of participants in group 1}

\item{n2}{number of participants in group 2}
}
\description{
variance of hedges' g\cr
equation #4.24 inboreinstein et al., 2009 (introduction to meta-analysis)
}
\examples{
varg(.3, 68)
dat \%>\% mutate (g = d.g(cohensd, totaln)) -> dat

}
