% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{z.r}
\alias{z.r}
\title{z to r}
\usage{
z.r(z)
}
\arguments{
\item{z}{fisher's z}
}
\description{
convert fisher's z to correlation coefficient
}
\examples{
single value
z.r(.3)

dplyr examples: dataset named "dat" with a column of correlation coefficient values named "fishz", create a new var called "r"
dat \%>\% mutate (r = z.r(fishz)) -> dat

}
