% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{zvar}
\alias{zvar}
\title{variance of fisher's z}
\usage{
zvar(n)
}
\arguments{
\item{n}{sample size from which the fisher's z is derived}
}
\description{
compute the sampling variance of fisher's z
}
\examples{
zvar(100)

}
