% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{dept.g}
\alias{dept.g}
\title{t to g}
\usage{
dept.g(t, n1, r, sdpre, sdpost, dir, k = 1)

dept.g(t, n1, r, sdpre, sdpost, dir, k = 1)
}
\arguments{
\item{t}{the paired-sample/dependent t-statistics}

\item{r}{the correlation between the paired scores}

\item{sdpre}{the standard deviation of the pre-test scores}

\item{sdpost}{the standard deviation of the post-test scores}

\item{dir}{empirical direction: +1 or -1. articles sometimes report t-statistics in absolute value, may be prudent to check when extracting stats\cr
if the reported t-statistics is positive, read text to ensure it reflects a positive association, enter the t as is and enter dir = 1 if so\cr
if the reported t-statistics is positive, but text suggests a negative association, enter the t as is and enter dir = -1\cr
if the reported t-statistics is negative, you can simply enter the negative t-value and enter dir = 1\cr
note: if your t-statistics is computed from taking the ratio of b and se, e.g., using bse.t(), then it should already have empirical direction "built-in" since b is directional itself, so simply enter dir = 1}

\item{n}{number of participants}
}
\description{
dependent (paired-samples) t-statistics to hedges' g\cr

dependent (paired-samples) t-statistics to cohen's d\cr
}
\examples{
dept.g(2.14, 20, 35, -1)
dat \%>\% mutate (g = dept.g(t_stat, n1, n2, direction)) -> dat

dept.d(2.14, 20, 35, -1)
dat \%>\% mutate (d = dept.g(t_stat, n1, n2, direction)) -> dat

}
