% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{meanci.r}
\alias{meanci.r}
\title{mean and ci to r}
\usage{
meanci.r(
  m1,
  m2,
  cil1,
  ciu1,
  cil2,
  ciu2,
  n1,
  n2,
  k,
  result = c("sediff", "cil", "ciu", "marge")
)
}
\arguments{
\item{m1}{mean of group 1}

\item{m2}{mean of group 2}

\item{cil1}{lower bound of 95 confidence interval}

\item{ciu1}{upper bound of 95 confidence interval}

\item{cil2}{lower bound of 95 confidence interval}

\item{ciu2}{upper bound of 95 confidence interval}

\item{n1}{cell size of group 1}

\item{n2}{cell size of group 2}

\item{k}{total number of predictors (k=0 for raw)}
}
\description{
means and confidence interval of two groups to correlation coefficient
}
\examples{
meanci.r(35, 28, 30, 40, 26, 30, 26, 54, 0, "ciu")
dat \%>\% mutate (r_from_meanci = meanci.r(stress_mean, control_mean, stress_cil, stress_ciu, control_cil, control_ciu, stress_n, control_n, num_predictors))-> dat

}
