% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{median.sd}
\alias{median.sd}
\title{median, minimum, and max to sd}
\usage{
\method{median}{sd}(median, min, max)
}
\arguments{
\item{median}{the median}

\item{min}{the minimum}

\item{max}{the maximum}
}
\description{
estimate sd from median, minimum and max. see \href{https://onlinelibrary.wiley.com/doi/full/10.4073/cmpn.2016.3}{polanin & snilstveit, 2016}
}
\examples{
median.sd(61, -11, 516)
dat \%>\% mutate (sd_from_medianminmax = median.sd(stress_median, stress_min, stress_max)) -> dat

}
