% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{se.sd}
\alias{se.sd}
\title{se to sd}
\usage{
se.sd(se, n)
}
\arguments{
\item{se}{standard error}

\item{n}{cell size}
}
\description{
standard error of raw mean to standard deviation
}
\examples{
se.sd(0.04, 15)

}
