% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{sediff}
\alias{sediff}
\title{differences in computed se}
\usage{
sediff(est, cil, ciu, n, k, type = c("reg", "exp b", "geo", "percent"))
}
\arguments{
\item{est}{the point estimate (e.g., a regression coefficient, a mean)}

\item{cil}{the lower bound of the 95 confidence interval}

\item{ciu}{the upper bound of the 95 confidence interval}

\item{n}{the sample size}

\item{k}{the total number of predictors. for converting the ci associated with a raw mean, enter k = 0.}

\item{type}{select one of:\itemize{
\item"reg" for regular coefficient/mean \cr
\item"exp b" for exponentiated regression coefficients (e.g., estimated odd ratios from exponentiated logistic regression coefficients) and cis \cr
\item"geo" for geometric means (exponentiated arithmetic means of logged raw units) and cis \cr
\item"percent" for percent changes (exponentiated regression coefficients expressed in percent changes) and cis
}}
}
\description{
the differences in the computed standard error using the upper bound vs. lower bound of 95 confidence interval\cr
computations are based on t-distribution, not z\cr
returns and prints a value/vector of difference(s)
}
\examples{
# e.g., exponentiated logistic regression coefficient (estimated odd ratios) and cis
se.diff(4.05, 1.56, 10.51, 141, 8, "exp b")

}
