% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaela.R
\name{vard}
\alias{vard}
\title{variance of d (independent samples)}
\usage{
vard(d, n1, n2)
}
\arguments{
\item{d}{cohen's d}

\item{n1}{number of participants in group 1}

\item{n2}{number of participants in group 2}
}
\description{
variance of cohen's d\cr
equation #4.20 inboreinstein et al., 2009 (introduction to meta-analysis)
}
\examples{
vard(.89, 24, 36)

}
